requirements.txt
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
pyspark>=3.4.0
scikit-surprise>=1.1.3
tqdm>=4.64.0
joblib>=1.2.0
jupyter>=1.0.0
seaborn>=0.12.0
